Implement SRP & DRY Priciples
