+++ backend/src/auth/auth.controller.ts
@@ -25,7 +25,7 @@ export class AuthController {
   @UseGuards(LocalAuthGuard)
   @Post('login')
   async login(@Request() req: { user: User }): Promise<any> {
-    console.log("hitted")
+    console.log('hitted');
     return this.authService.login(req.user);
   }
